Added PEP2 scraper
Replaced old scrape.js
